refactor: layout 단의 로직 훅으로 분리
