+++ src/hooks/@common/useModal.ts
@@ -1,0 +1,15 @@
+import { useState } from 'react';
+
+const useModal = () => {
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const handleModalOpen = () => {
+    setIsModalOpen(true);
+  };
+  const handleModalClose = () => {
+    setIsModalOpen(false);
+  };
+
+  return { isModalOpen, handleModalOpen, handleModalClose };
+};
+
+export default useModal;
